Updates for name change and fixes to tasks.py
