+++ docs/gallery.rst
@@ -4,9 +4,9 @@ Gallery
 
 .. todo::
 
-   This section will showcase various use cases for optiwrap
+   This section will showcase various use cases for boa
 
-Projects using optiwrap
+Projects using boa
 -----------------------
 
 - `FETCH3 <https://fetch3-nhl.readthedocs.io/en/latest/index.html>`_
