+++ docs/overview.rst
@@ -3,7 +3,7 @@ Overview
 ########
 
 .. todo::
-   - General description of how optiwrap works
+   - General description of how boa works
    - Overview of package structure
    - Overview of basic steps for creating a model wrapper
    - Overview of main features
