--- optiwrap/__init__.py
@@ -1,11 +1,0 @@
-"""
-Optiwrap package
-"""
-__version__ = "0.4.7"
-
-from optiwrap.ax_instantiation_utils import *  # noqa
-from optiwrap.metrics.metrics import *  # noqa
-from optiwrap.runner import WrappedJobRunner
-from optiwrap.storage import *  # noqa
-from optiwrap.wrapper import BaseWrapper
-from optiwrap.wrapper_utils import *  # noqa
