+++ tasks.py
@@ -70,7 +70,7 @@ All Style Checks Passed Successfully
 
 
 @task
-def pytest(command, options):
+def pytest(command, options=""):
     """Runs pytest to identify failing tests and doctests"""
 
     print(
@@ -93,9 +93,9 @@ Running Sphinx to test the docs building
 ========================================
 """
     )
-    shutil.rmtree("docs/_build")
-    shutil.rmtree("docs/api")
-    shutil.rmtree("docs/jupyter_execute")
+    shutil.rmtree("docs/_build", ignore_errors=True)
+    shutil.rmtree("docs/api", ignore_errors=True)
+    shutil.rmtree("docs/jupyter_execute", ignore_errors=True)
     command.run("sphinx-build -b html docs docs/_build", echo=True, pty=POSIX)
